+++ src/controllers/parkingspot.js
@@ -9,6 +9,8 @@ module.exports = {
   //PARKING SPOTS
   
   createParkingSpot: (cameraId, spotId, callback) => {
+    if(!cameraId || !spotId)
+      return callback(new Error("Wrong parameters"), null)
     attemptConnection(`INSERT INTO parkingspot VALUES(${mysql.escape(cameraId)},${mysql.escape(spotId)},FALSE)`, (err,res) => {
       if (err) throw err
       else {
@@ -19,6 +21,8 @@ module.exports = {
     })
   },
   getParkingSpotInfos: (cameraId, spotId, callback) => {
+    if(!cameraId || !spotId)
+      return callback(new Error("Wrong parameters"), null)
     attemptConnection(`SELECT IsTaken FROM parkingspot WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) throw err
       else {
@@ -33,6 +37,8 @@ module.exports = {
     })
   },
   updateParkingSpot: (cameraId, spotId, state, callback) => {
+    if(!cameraId || !spotId || !state)
+      return callback(new Error("Wrong parameters"), null)
     attemptConnection(`UPDATE parkingspot SET IsTaken = ${mysql.escape(state)} WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) throw err
       else {
@@ -42,6 +48,8 @@ module.exports = {
     })
   },
   deleteParkingSpot: (cameraId, spotId, callback) => {
+    if(!cameraId || !spotId)
+      return callback(new Error("Wrong parameters"), null)
     attemptConnection(`DELETE FROM parkingspot WHERE CameraId = `+mysql.escape(cameraId)+` AND SpotNumber = `+mysql.escape(spotId), (err,res) => {
       if (err) throw err
       else {
